// sso/principal-assignments/ssoadmin/ssoadmin.d.ts
  export interface CreateAccountAssignmentRequest {
    /**
     * The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.
     */
    InstanceArn: InstanceArn;
    /**
     * The identifier for the chosen target.
     */
    TargetId: TargetId;
    /**
     * The entity type for which the assignment will be created.
     */
    TargetType: TargetType;
    /**
     * The ARN of the permission set that the admin wants to grant the principal access to.
     */
    PermissionSetArn: PermissionSetArn;
    /**
     * The entity type for which the assignment will be created.
     */
    PrincipalType: PrincipalType;
    /**
     * The identifier of the principal.
     */
    PrincipalId: PrincipalId;
  }
// ...
  export interface DeleteAccountAssignmentRequest {
    /**
     * The ARN of the SSO instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.
     */
    InstanceArn: InstanceArn;
    /**
     * The identifier for the chosen target.
     */
    TargetId: TargetId;
    /**
     * The entity type for which the assignment will be deleted.
     */
    TargetType: TargetType;
    /**
     * The ARN of the permission set that will be used to remove access.
     */
    PermissionSetArn: PermissionSetArn;
    /**
     * The entity type for which the assignment will be deleted.
     */
    PrincipalType: PrincipalType;
    /**
     * The identifier of the principal.
     */
    PrincipalId: PrincipalId;
  }
// ...
  export type InstanceArn = string;
// ...
  export type PermissionSetArn = string;
// ...
  export type PrincipalId = string;
  export type PrincipalType = "USER"|"GROUP"|string;
// ...
  export type TargetId = string;
  export type TargetType = "AWS_ACCOUNT"|string;